Renamed ControlPointV0005 to ControlPointV0003
